requirements.txt
django
djangorestframework
environs
openai
langchain
pydantic